custom launcher for debugging

+++ ServerLauncher/Form1.cs
@@ -26,67 +26,74 @@ namespace ServerLauncher
 
         private void B_start_Click(object sender, EventArgs e)
         {
-            AccountCacher = new Process();
-            AccountCacher.StartInfo.FileName = "AccountCacher.exe";
-            AccountCacher.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            AccountCacher.Start();
+            B_start.Enabled = false;
 
-            Thread.Sleep(500);
+            if (StartAccountCheckBox.Checked)
+            {
+                AccountCacher = new Process();
+                AccountCacher.StartInfo.FileName = "AccountCacher.exe";
+                AccountCacher.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                AccountCacher.Start();
 
-            LauncherServer = new Process();
-            LauncherServer.StartInfo.FileName = "LauncherServer.exe";
-            LauncherServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            LauncherServer.Start();
+                Thread.Sleep(500);
+            }
 
-            Thread.Sleep(500);
+            if (StartLauncherCheckBox.Checked)
+            {
+                LauncherServer = new Process();
+                LauncherServer.StartInfo.FileName = "LauncherServer.exe";
+                LauncherServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                LauncherServer.Start();
 
-            LobbyServer = new Process();
-            LobbyServer.StartInfo.FileName = "LobbyServer.exe";
-            LobbyServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            LobbyServer.Start();
+                Thread.Sleep(500);
+            }
 
-            Thread.Sleep(500);
+            if (StartLobbyCheckBox.Checked)
+            {
+                LobbyServer = new Process();
+                LobbyServer.StartInfo.FileName = "LobbyServer.exe";
+                LobbyServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                LobbyServer.Start();
 
-            WorldServer = new Process();
-            WorldServer.StartInfo.FileName = "WorldServer.exe";
-            WorldServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
-            WorldServer.Start();
+                Thread.Sleep(500);
+            }
+
+            if (StartWorldCheckBox.Checked)
+            {
+                WorldServer = new Process();
+                WorldServer.StartInfo.FileName = "WorldServer.exe";
+                WorldServer.StartInfo.WindowStyle = ProcessWindowStyle.Minimized;
+                WorldServer.Start();
+            }
+
+            B_start.Enabled = true;
         }
 
         private void B_stop_Click(object sender, EventArgs e)
         {
             try
             {
-                WorldServer.Kill();
+                if( WorldServer != null ) WorldServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                LobbyServer.Kill();
+                if (LobbyServer != null) LobbyServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                LauncherServer.Kill();
+                if (LauncherServer != null) LauncherServer.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
+            
             try
             {
-                AccountCacher.Kill();
+                if (AccountCacher != null) AccountCacher.Kill();
             }
-            catch (Exception)
-            {
-
-            }
+            catch (Exception) { }
         }
     }
 }
